Demos/AI-3002/mslearn-ai-info-extraction-AI-3002-Develop-AI-information-extraction-solutions-in-Azure-AI-102-Day-5-main/Labfiles/prebuilt-doc-intelligence/Python/requirements.txt
python-dotenvazure-identity
azure-ai-documentintelligence